Remove unused import and update comments in httpclient/request.go
