+++ httpclient/request.go
@@ -8,7 +8,6 @@ import (
 	"time"
 
 	"github.com/deploymenttheory/go-api-http-client/authenticationhandler"
-	"github.com/deploymenttheory/go-api-http-client/cookiejar"
 	"github.com/deploymenttheory/go-api-http-client/headers"
 	"github.com/deploymenttheory/go-api-http-client/httpmethod"
 	"github.com/deploymenttheory/go-api-http-client/logger"
@@ -154,7 +153,7 @@ func (c *Client) executeRequestWithRetries(method, endpoint string, body, out in
 	c.ConcurrencyHandler.Metrics.TotalRequests++
 	c.ConcurrencyHandler.Metrics.Lock.Unlock()
 
-	// Perform Request
+	// Create a new HTTP request with the provided method, URL, and body
 	req, err := http.NewRequest(method, url, bytes.NewBuffer(requestData))
 	if err != nil {
 		return nil, err
@@ -315,7 +314,7 @@ func (c *Client) executeRequest(method, endpoint string, body, out interface{}) 
 	// Construct URL using the ConstructAPIResourceEndpoint function
 	url := c.APIHandler.ConstructAPIResourceEndpoint(c.InstanceName, endpoint, log)
 
-	// Perform Request
+	// Create a new HTTP request with the provided method, URL, and body
 	req, err := http.NewRequest(method, url, bytes.NewBuffer(requestData))
 	if err != nil {
 		return nil, err
@@ -337,11 +336,8 @@ func (c *Client) executeRequest(method, endpoint string, body, out interface{}) 
 		return nil, err
 	}
 
-	// Get Cookies
-	cookiejar.GetCookies(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		// Log outgoing cookies
-		log.LogCookies("incoming", r, method, endpoint)
-	}), c.Logger).ServeHTTP(nil, req)
+	// Log outgoing cookies
+	log.LogCookies("incoming", req, method, endpoint)
 
 	// Checks for the presence of a deprecation header in the HTTP response and logs if found.
 	headers.CheckDeprecationHeader(resp, log)
@@ -456,97 +452,4 @@ func (c *Client) handleSuccessResponse(resp *http.Response, out interface{}, log
 		zap.Int("status_code", resp.StatusCode),
 	)
 	return nil
-}
-
-// DoMultipartRequest creates and executes a multipart HTTP request. It is used for sending files
-// and form fields in a single request. This method handles the construction of the multipart
-// message body, setting the appropriate headers, and sending the request to the given endpoint.
-//
-// Parameters:
-// - method: The HTTP method to use (e.g., POST, PUT).
-// - endpoint: The API endpoint to which the request will be sent.
-// - fields: A map of form fields and their values to include in the multipart message.
-// - files: A map of file field names to file paths that will be included as file attachments.
-// - out: A pointer to a variable where the unmarshaled response will be stored.
-//
-// Returns:
-// - A pointer to the http.Response received from the server.
-// - An error if the request could not be sent or the response could not be processed.
-//
-// The function first validates the authentication token, then constructs the multipart
-// request body based on the provided fields and files. It then constructs the full URL for
-// the request, sets the required headers (including Authorization and Content-Type), and
-// sends the request.
-//
-// If debug mode is enabled, the function logs all the request headers before sending the request.
-// After the request is sent, the function checks the response status code. If the response is
-// not within the success range (200-299), it logs an error and returns the response and an error.
-// If the response is successful, it attempts to unmarshal the response body into the 'out' parameter.
-//
-// Note:
-// The caller should handle closing the response body when successful.
-func (c *Client) DoMultipartRequest(method, endpoint string, fields map[string]string, files map[string]string, out interface{}) (*http.Response, error) {
-	log := c.Logger
-
-	// Auth Token validation check
-	// valid, err := c.ValidAuthTokenCheck()
-	// if err != nil || !valid {
-	// 	return nil, err
-	//}
-
-	// Auth Token validation check
-	clientCredentials := authenticationhandler.ClientCredentials{
-		Username:     c.clientConfig.Auth.Username,
-		Password:     c.clientConfig.Auth.Password,
-		ClientID:     c.clientConfig.Auth.ClientID,
-		ClientSecret: c.clientConfig.Auth.ClientSecret,
-	}
-
-	valid, err := c.AuthTokenHandler.ValidAuthTokenCheck(c.APIHandler, c.httpClient, clientCredentials, c.clientConfig.ClientOptions.TokenRefreshBufferPeriod)
-	if err != nil || !valid {
-		return nil, err
-	}
-
-	// Determine which set of encoding and content-type request rules to use
-	//apiHandler := c.APIHandler
-
-	// Marshal the multipart form data
-	requestData, contentType, err := c.APIHandler.MarshalMultipartRequest(fields, files, log)
-	if err != nil {
-		return nil, err
-	}
-
-	// Construct URL using the ConstructAPIResourceEndpoint function
-	url := c.APIHandler.ConstructAPIResourceEndpoint(c.InstanceName, endpoint, log)
-
-	// Create the request
-	req, err := http.NewRequest(method, url, bytes.NewBuffer(requestData))
-	if err != nil {
-		return nil, err
-	}
-
-	// Initialize HeaderManager
-	//log.Debug("Setting Authorization header with token", zap.String("Token", c.Token))
-	headerHandler := headers.NewHeaderHandler(req, c.Logger, c.APIHandler, c.AuthTokenHandler)
-
-	// Use HeaderManager to set headers
-	headerHandler.SetContentType(contentType)
-	headerHandler.SetRequestHeaders(endpoint)
-	headerHandler.LogHeaders(c.clientConfig.ClientOptions.HideSensitiveData)
-
-	// Execute the request
-	resp, err := c.do(req, log, method, endpoint)
-	if err != nil {
-		return nil, err
-	}
-
-	// Check for successful status code
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		// Handle error responses
-		//return nil, c.handleErrorResponse(resp, log, "Failed to process the HTTP request", method, endpoint)
-		return nil, c.handleErrorResponse(resp, out, log, method, endpoint)
-	} else {
-		// Handle successful responses
-		return resp, c.handleSuccessResponse(resp, out, log, method, endpoint)
-	}
 }
